unit tests fixes

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/state/derived/MinMaxValue.java
@@ -4,14 +4,17 @@
 
 package com.apollocurrency.aplwallet.apl.core.dao.state.derived;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public class MinMaxValue {
-    private long min;
-    private long max;
+    private BigDecimal min;
+    private BigDecimal max;
     private String column;
     private long count;
     private int height;
